chore: use different tsconfig frontend and node
