+++ package.json
@@ -10,9 +10,9 @@
     "prepare-demo-app": "npm install --prefix demo-app && npm run build --prefix demo-app",
     "react:start": "react-scripts start",
     "react:build": "react-scripts build",
-    "electron:copy": "cp -r ./electron/. ./build",
-    "dev": "concurrently -k \"BROWSER=none npm run react:start\" \"tsc -w\" \"wait-on tcp:3000 && nodemon --exec electron build/electron/electron.js\"",
-    "build": "npm run clean && npm run react:build && tsc",
+    "dev": "concurrently -k \"BROWSER=none npm run react:start\" \"npm run tsc-electron -- -w\" \"wait-on tcp:3000 && nodemon --exec electron build/electron/electron.js\"",
+    "build": "npm run clean && npm run react:build && npm run tsc-electron",
+    "tsc-electron": "tsc -p tsconfig.node.json",
     "pack": "npm run build && electron-builder",
     "release": "npm run pack -- -mwl --publish=always",
     "fix-sharp": "node node_modules/sharp/install/libvips && node node_modules/sharp/install/dll-copy",
