+++ src/components/ExportScriptFlyout/Flyout.tsx
@@ -26,7 +26,8 @@ import { EuiFlyout } from '@elastic/eui';
 import type { Steps } from '@elastic/synthetics';
 import React, { useContext, useEffect, useMemo, useState } from 'react';
 import { getCodeFromActions } from '../../common/shared';
-import type { JourneyType, Setter } from '../../common/types';
+import type { Setter } from '../../common/types';
+import type { JourneyType } from '../../../shared/types';
 import { CommunicationContext } from '../../contexts/CommunicationContext';
 import { Body } from './Body';
 import { Footer } from './Footer';
