+++ src/hooks/useSyntheticsTest.ts
@@ -27,7 +27,7 @@ import { IpcRendererEvent } from 'electron';
 import { useCallback, useContext, useEffect, useReducer, useState } from 'react';
 import { getCodeFromActions, getCodeForFailedResult } from '../common/shared';
 import { CommunicationContext } from '../contexts/CommunicationContext';
-import type { Result, TestEvent } from '../common/types';
+import type { Result, TestEvent } from '../../shared/types';
 import type { ITestContext } from '../contexts/TestContext';
 import { resultReducer } from '../helpers/resultReducer';
