+++ tsconfig.json
@@ -1,22 +1,14 @@
+/**
+  tsconfig for frontend. It is supposed to be `tsconfig.frontend.json`
+  but react-scripts can only understand `tsconfig.json`(cannot specify tsconfig path)
+*/
+
 {
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "target": "ES2021",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "noFallthroughCasesInSwitch": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "incremental": true,
+    "module": "esnext",
+    "noEmit": true,
     "jsx": "react-jsx",
-    "isolatedModules": true,
-    "outDir": "build"
   },
-  "include": ["src/**/*", "electron/**/*"],
-  "exclude": ["node_modules"]
+  "include": ["src/**/*"]
 }
