+++ tsconfig.node.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "module": "commonjs",
+  },
+  "include": ["electron/**/*"],
+}
